Make the AWS region, AMI, keypair, instance type and OpenShift binary customisable
The AWS region, AMI, keypair and instance type are not hardcoded anymore.
They can be customized in the config file.
If the config file is empty, the default values are the previous hardcoded ones.
It is also possible to specify a local OpenShift binary to deploy.
If not specified, it will intall one from the yum repository.