Fix VxLAN creation
When we create the VxLAN links between minions, we have to loop over all the minions
and to skip ourself. A minion doesn’t need a link to itself.

The "groups['tag_env-host-type_' + oo_env + '-openshift-minion']" over which we iterate contains the public IPs.
So, we have to use facter_ec2_public_ipv4 to get the public IP of the current minion when doing the filtering.