Re-include config.yml from launch.yml but make it optional
It is not possible to first launch and configure the master and then to launch
and configure the minions because the master needs to know the IPs of the
minions (in the --nodes parameter found in /etc/sysconfig/openshift)

If we proceed in this order:
* launch the master machine
* configure the master
* launch the minion machines
* configure the minions
what happens is that the "--nodes" parameter of openshift is empty.
This prevents openshift master from starting.
Then the minions are starting, but they fail because they were enable to
connect to the master.
At the end, neither openmaster nor openshift minions are running.

In order to have things properly configured, i.e. to have the master know
the minions IPs and the minions knows the master IP, we need to proceed in
this order:
* launch the master machine
* launch the minion machines
* configure the master
* configure the minion

The "config" tag is used to skip the configuration step when launching
the machines. cluster.sh already takes care of doing the "config" step
after the "launch" one.