Adding support for labeling nodes as unscheduleable
For idempotency it first checks the output of 'oc get node <hostname>' to see
if any action needs to be taken.  The trick was waiting to make sure that the
node autoregistration had actually happened.

If you set openshift_scheduleable=False in the inventory then the node will be
marked as unscheduleable.  Likewise setting it to True undoes the change.
Having openshift_scheduleable undefined is the same as True since that is the
default state of a Node.