Fix exception handling.
The subcommand of the action was called using os.system. The exit value
of os.system is a 16-bit value. This value was propagated and used as
exit value of the whole `cluster {ACTION}` command without any
modification, resulting in `exit()` being called with value > 255. In
the CPython 2.7 exit(v) with v > 255 behaves like exit(0), which hides
that we had an error during the execution.

This commit removes the error propagation by return value and introduces
using exceptions instead.