Fix bug with not upgrading openshift-master to atomic-openshift-master.
Removing the full call to config resulted in rpms not getting upgraded. Config
was doing a yum update of everything, which picks up the
atomic-openshift-master obsoleting openshift-master. The actual yum call
changed here would not. Instead we switch to a direct call to yum which
correctly picks up the obsoletes and updates to atomic-openshift packages.