Add pre-upgrade script to be run on first master.
Script currently just checks for port names that are no longer valid. In theory
other checks may be added to this script in the future. Script was originally
written by Steve Milner and Andy Goldstein.

If the script fails, ansible seems to handle this nicely by default, exiting
the upgrade and displaying stderr and stdout, both of which contain useful info
on what the problem was.