Fix missing etcd_data_dir bug.
A late change to original PR was not properly tested, there is a problem in the
facts when upgrading where the deployment type is openshift-enterprise, and the
system facts start reporting data_dir and config_base as referencing origin
directories, which are not yet symlinked to their previous openshift variants.

To correct we watch for a scenario where we evaluate these to origin
directories, which don't exist, but the openshift ones do. (to allow for
installation to still point at the origin variety)