Handling preconfigured load balancers
The preconfigured load balancers, previously denoted by having 'run_on' set to
false, cannot have their facts gathered which results in a stack trace.  Later
when we write out the inventory we have to fake out the hostname and just use
'connect_to'.

We're likely going to have the concept of other types of "plug-in" hosts where
we don't run ansible.  We should make sure we abstract this properly so it's
easy to add additional types of hosts.

Also in the commit:
- Renamed 'run_on' to 'preconfigured' and inverted the logic as needed
- Output tally of Masters and Nodes as well as remaining Masters required for
  HA
- Minor rewording in a few places
- Currently only prompting for the load balancer after all other hosts have
  been entered
- Removed spurious echo