Block re-use of master/node as load balancer in attended install.
Code was present to catch this in unattended installs but was looking for a
host record with both master/node and master_lb set to true, but in the
attended installs we were adding a separate host record with the same
connect_to.

Attended tests can now optionally specify multiple "attempted" strings for the
master_lb specification, we'll try to input each if multiple are specified.

Cleanup some empty defaults and error messages as well.