Adjust requirement for 3 masters for HA deployments.
If only 2 masters are specified, consider this a configuration error if running
an unattended install, and prevent it completely if running an attended
install. (continues to prompt for hosts until you have at least 3)

Because this condition cannot be entered in the interactive install, we can't
really write a test for this negative case.