a-o-i: Redo logic for detecting master_lb
After much research and debugging, it turns out the facts
for nativeha hosts were much less consistent than I hoped.
With this newfound knowledge, I've implemented this method.

In short, you check with the first master to see what it thinks
the master_lb is, then you include that host as "installed".