Support openshift_image_tag
This is the containerized openshift_pkg_version equivalent.  Originally I was
hoping to reuse openshift_pkg_version for containerized installs but the fact
that it's very coupled to yum made that pretty ugly.

However, I did opt to rely on the previously existing 'openshift_version'
variable.  Containerized and RPM installs can both use that variable and it
will be set appropriately if either openshift_pkg_version or
openshift_image_tag are set.  I suspect someday containerized installs will be
the only option and I didn't can to have thinkgs like openshift_pkg_version and
openshift_image_tag in the playbooks anymore the necessary.