OSE/Origin < 3.2/1.2 should not get Docker 1.9
RHEL and Centos docker-1.9 RPMs will have a Conflicts for origin and
atomic-openshift < 3.2. If we left yum to its own devices we'd get a mess when
trying to install once these packages land in the repos. So avoid installing
docker-1.9 on Origin/OSE < 1.2/3.2.
See: https://bugzilla.redhat.com/show_bug.cgi?id=1304038

Fedora doesn't keep old packages in repos so we just get the latest there.
Docker 1.10 should be fine if and when it becomes available.