Fetching the current version a little more carefully
Previously we were trying to use the running container to get the current
version.  There are cases in which the Master or Node may not be running during
upgrade.  It's actually safer to just run the container to fetch the version
that would be launch if the container were running.  Then we pull the image to
see what the latest image contains.