Fix issues when docker is not pre-installed.
A docker run command was recently guarded to only be run when docker is
installed, this change does the same for an additional docker exec.

Adjusted some logic that might make for inaccurate openshift version
reporting when containerized is true, but for some reason openshift is
installed on the main OS.

Fixed an additional failure where a recent fix to reload the docker
service would run after the docker rpm was just installed, and the
service is not yet loaded. Due to it's nature as a workaround, it does
not look like it can be moved to a better location, so ignoring errors
appears to be the best solution.