Remove /dev mount for node container
This breaks some storage plugins but it also breaks non root logins. A fix for
docker is coming in a future updated to docker-1.9.1 at which time we'll revert
this once we have a method to ensure we're running a fixed version.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1329326