Fix openshift_generate_no_proxy_hosts boolean
Fixing this also made it obvious that we weren't adding even the current host's
name to the no_proxy list. This is absolutely necessary or the master won't be
able to reach etcd. So even if they request not to have the list of all hosts
and cluster dns zone added we should add the current host's hostname to the
no_proxy list.