Fixed openvswitch not upgrading.
Problem was a missing restart as systemd file did not change, so no
notification was triggered. Added a notification on a file that
contains the image version which does change.

This exposed a bug where openvswitch shuts down the node and you return
to no functioning nodes. Fixed by reordering the handlers so openvswitch
restarts first.