Add new registries during upgrade playbook.
The situation with upgrades and pre.yml vs upgrade.yml is a bit
convoluted. Previously we were running some pre tasks, which included
updating the systemd units, which have automatic notifiers to restart
the services, and would attempt to pull down the new images immediately.
This should not be happening in pre.yml, but it also meant that we could
not actually pull the images if we're adding a new registry for this run
of the upgrade playbook. (i.e. you would have to re-run config playbook
first to get the new registry added)

This change drops the use of systemd unit file updates in pre.yml
completely, they were already re-applied in the actual upgrade plays.

We also stop using pre_tasks and then openshift_cli (which modifies the
wrappers around openshift tooling, i.e. a write operation) in pre.yml.

Instead we use the openshift_docker role which is a little safer in this
regard, and will update /etc/sysconfig/docker so we get our additional
registries and can safely lookup the new version.

openshift_cli role is now run early in the actual upgrade plays to
ensure the wrappers are updated appropriately.