Fixes for openshift_docker_hosted_registry_insecure var.
Fixes a failure on masters if you explicitly set
openshift_docker_hosted_registry_insecure=true. This is the default but
if you tried to set it an error would trigger as a relevant variable was
not passed in the master playbooks.

Fixes setting the variable to false being ignored.
master/node playbooks were referencing the docker fact, which was not
set at that point and thus we were always getting the default of true,
regardless what was in your inventory.

Stop passing registry insecure in via playbooks, we can access it when running
openshift_facts itself. Add a new default in openshift facts.