Refactor openshift_version behavior.
Very early in playbooks we must init the openshift_version for each
host. First we determine it for the master, logic now is pushed into the
openshift_docker role which we run only on first master via
openshift_cli. Facts are reloaded leaving us with a first master with
openshift.common.version fact we can then re-use on all other hosts. The
correct version of docker should be installed as well.

We then set openshift_version for all other hosts by re-using the master
fact.