Attempt to fix containerized node start failure with Docker 1.10.
It appears that in some situations (can't reliably reproduce yet), node
will fail to start. This appears to be related to the node-dep service
and possibly it's environment file.

This file is also an EnvironmentFile for the node service, but it's only
created by the node-dep service, and it looks like it may try to read
it's environment before the node-dep service has fully started and
created the file.

Workaround with a an explicit service start.