Update docker upgrade playbook to be more flexible.
Now more of a generic upgrade playbook to go to the latest Docker
version.

Added support for docker_version inventory variable, in which case we
disable the check for >= 1.10 and make sure you're running at least the
specified version. (we will not downgrade you to the requested version
however, this is much too complicated)