Copy openshift binaries instead of using wrapper script.
For containerized masters, we previously create a wrapper script to run
the docker CLI image and clean up afterward, but this approach is much
slower than having the binary on the native system.

Instead we now use an ansible module to handle the logic of syncing the
various OpenShift binaries and symlinks for client tools out onto the
host. The module will correctly report changed if modifications were
needed.

Substantial speed improvement for containerized installs which requires
many openshift/oc commands.