Fixed a bug in modify_yaml module.
When adding a property to a section that does not yet exist, the
property was being inserted into the section above where it should have
due to the pointer not being adjusted for the new section we created.

Adds a test sub-directory, would like to eventually move all the
ooinstall tests here as well so nosetests just works from top level
directory.