a-o-i: Better inventory group handling
A more dynamic and flexible method of writing out host groups to the ansible
inventory file. To accompany this, in the quick-installer interactive mode
the etcd role is automatically applied to any masters. Anyone hand writing
the quick installer config will now have to explicitly specify which hosts
will be etcd.

Fixes #2200