Open OpenStack security group for the service node port range
With OpenShift 3.2, creating a service accessible from the outside of the
cluster thanks to `nodePort` automatically opens the “local” `iptables`
firewall to allow incoming connection on the `nodePort` of the service.

In order to benefit from this improvement, the OpenStack security group
shouldn’t block those incoming connections.

This change opens, on the OS nodes, the port range dedicated to service
node ports.