Stop restarting node after upgrading master rpms.
This looks to be causing a customer issue where some HA upgrades fail,
due to a missing EgressNetworkPolicy API. We update master rpms, we
don't restart services yet, but then restart node service which tries to
talk to an API that does not yet exist. (pending restart)

Restarting node here is very out of place and appears to not be
required.