Fix iptables dependency problems.
In situations where iptables is already installed but is not the latest
version, and iptables-services is *not* installed, this task would fail
as ansible tries to install the newest services rpm, but it requires a
change to iptables version. (which is not handled gracefully) This
situation will very likely trigger on the release of 7.3.

Instead, switch to installing "latest" versions of each package, and
handle in two separate tasks so the iptables rpm is already latest
before we proceed to services.

Fixes #1389722