Added a BYO playbook for configuring NetworkManager on nodes
In order to do a full install of OpenShfit using the byo/config.yml
playbook, it is currently required that NetworkManager be installed
and configured on the nodes prior to the installation. This playbook
introduces a very simple default configuration that can be used to
install, configure and enable NetworkManager on their nodes.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>