Check for bad versions of yum and subscription-manager.
Use of yum and repoquery will output the given additional warning when
using newer versions of subscription-manager, with older versions of
yum. (RHEL 7.1) Installing/upgrading newer docker can pull this
subscription-manager in resulting in problems with older versions of
ansible and it's yum module, as well as any use of repoquery/yum
commands in our playbooks.

This change explicitly checks for the problem by using repoquery and
fails early if found. This is run early in both config and upgrade.