Simplify OpenShift service debug levels.
Stop using facts for debug levels entirely as this is not a fact about
the system, rather a configuration parameter of the playbooks/inventory.

Deprecate debug_level and rename to openshift_debug_level. Default is
2 if undefined, and can be specified separately for master/node if
desired.

Moved all definitions/defaulting straight to systemd_units.yml for
master/node where the values are actually used. Here we check if
openshift_debug_level is set and default it if not, and then set the
master/node variant if they are not set.

NOTE: Vars can't be used in the roles here as the systemd_units.yml
files are used separtely during upgrade.