Fix the list done after cluster creation on libvirt and OpenStack
The `list.yml` playbooks are using cloud provider specific variables to find
the IPs of the VMs since 82449c6.

Those “cloud provider specific” variables are the ones provided by the dynamic
inventories.

But there was a problem when the `list.yml` playbooks are invoked from the
`launch.yml` ones because, in that case, the inventory is not coming from the
dynamic inventory scripts, but from the `add_host` done inside
`launch_instances.yml`.

Whereas the GCE and AWS `launch_instances.yml` were correctly adding in the
`add_host` the variables used by `list.yml`, libvirt and OpenStack were missing
that.

Fixes #2856