Speed up 'make ci' and trim the output
The virtualenv is conditionally rebuilt now if the test requirements
file is modified. This will save upwards of 30+ seconds in iterative
'make ci' runs.

The pylint output is now trimmed to exclude disabled tests.

The order of the 'ci' target prerequisites has been changed to run the
fastest tests first.

Closes #2933