Wait for nodes to be ready before proceeding with upgrade.
Near the end of node upgrade, we now wait for the node to report Ready
before marking it schedulable again. This should help eliminate delays
when pods need to relocate as the next node in line is evacuated.

Happens near the end of the process, the only remaining task would be to
mark it schedulable again so easy for admins to detect and recover from.