Validate system restart policy during pre-upgrade.
This was done far into the process potentially leaving the user in a
difficult situation if they had now considered they were running the
upgrade playbook on a host that would be restarted. Instead check
configuration and what host we're running on in pre-upgrade and allow
the user to abort before making any substantial changes.

This is a step towards merging master upgrade into one serial process.