Remove spurious argument
The script nuke_images.sh was introduced in
0c31d72be3bf32f848eedad9859a81ba858f8c8f and seems that the shell
argument $1 was never used (the only $1 in the script back then refers
to a field in a awk script).
There was a reference to $2 that was always undefined/empty.

The script was then simplified in
b89c835e3235f2628b37de15713c311d1b5a4bad, removing any reference to $1
and $2.

This commit cleans up the only call site to the shell script.