More reliable wait for master after full host reboot.
This would fail in some environments where openshift.common.hostname was
not reachable from the local system running ansible. Additionally it is
also possible the API port would not be available even if the host were.

Instead we wait for master ssh connection to come back up, then wait for
the master API on the remote host just as we do for the "service" style
restart.