facts/main: Require Python 3 for Fedora, Python 2 everywhere else
For a few reasons; among them that we currently have a dependency on `PyYAML`
which on Fedora Atomic Host isn't installed for Python 2 by default. Further,
many dependencies are being ported in Fedora to be Python 3.

Conversely, ensure that we're using Python 2 everywhere else (which is really
CentOS/RHEL), since AFAIK we don't test that path at all, and we really don't
need *more* entries in the supported matrix.