preflight checks: refactor and fix aos_version
Bring in openshift_repos to provide Origin repo before checks on Origin.

For OCP we want the check to fail if both version 3.3 and version 3.4
are available - they shouldn't have both channels enabled.
For Origin everything is in one repo so it's not surprising to find 1.4
and 1.5 versions available.

Added unit tests as well.