vendor patched upstream docker_container module.
Due to the use of a restricted name in the core `docker_container`
module's result, any standard output of a docker container captured
in the module's response was stripped out by ansible.

Because of this, we are forced to vendor a patched version of this
module, until a new version of ansible is released containing the
patched module.

This file should be removed once we begin requiring a release of ansible
containing the patched `docker_container` module.

This patch was taken directly from upstream, with no further changes:
20bf02f6b96356ab5fe68578a3af9462b4ca42a5