Install Ansible plugins into the 2.0 locations without clobber
A plugin for Ansible 1.9 was expected to reside under a different path,
`/usr/share/ansible_plugins/<type>_plugins/<name>.py`, than it is expected
to in Ansible 2.0, `/usr/share/ansible/plugins/<type>/<name>.py`. When a
user has plugins already installed to these locations for Ansible 2.0,
it is currently impossible to install `atomic-openshift-utils` as this
RPM places plugins in the old location and creates the new locations as
symlinks. As this project no longer supports Ansible 1.9, it is possible
to install plugins into the new locations for Ansible 2.0 and do so
without clobbering pre-existing plugins the user may have installed.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>