Do not require python-six via openshift_facts
It turns out the code in
roles/openshift_health_checker/openshift_checks/__init__.py had a
mistake in the import statement, and not the case that `add_metaclass`
is not available in the embedded `six` from `ansible.module_utils`.

The fix is to correct the import and do not require python-six to be
installed on target hosts.

Note: now that we have some unit tests in the role, errors like that
would have been caught.