Fix private CA
When private (aka not publicly trusted) CAs are used, and a registry from outside the cluster itself is used, it is not possible to retrieve images from it (without this fix).

This is only a problem of the containerized version, since in the RPM version, these files are loaded by design.

I think that also this adds no possible threats, since we are adding to the master-api containers CAs that are already trusted by the nodes, which have a higher security criticality.