repo: start testing PRs on Fedora Atomic Host
This patch adds a YAML file to hook up the repo to redhat-ci:
https://github.com/jlebon/redhat-ci

It is much like Travis CI: a YAML file describes the testing environment
as well as the tests themselves. However, it is more powerful than
Travis CI in terms of what it can provision. (For example, in this
patch, we're running two testsuites, each of them provisioning three
nodes and one container). For more details on supported fields, see:

https://github.com/jlebon/redhat-ci/blob/master/sample.redhat-ci.yml

The tests we add here simply check that the installer runs successfully
on Fedora Atomic Host and that the resulting cluster passes some basic
conformance tests. We target both v1.5.0-rc.0 and v3.6.0-alpha.0.