Cast etcd_debug to a boolean
In the current implementation, any user-provided string in the
`etcd_debug` variable will be placed into `etcd.conf`. The YAML
and Ansible boolean parsing is more generous than the Golang one,
so valid YAML booleans like `no` will be invalid when passed to
etcd. Casting to a boolean before casting to a string normalizes
the field.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>