Add /etc/sysconfig/etcd to etcd_container
The `/etc/sysconfig/etcd` file is often used to allow the usage of additional variables passed to the etcd daemon.
Example of this is to make some new nodes join an existing cluster.

Having two files simplify the automation around it, since in the `/etc/sysconfig/etcd` is possible to put temporary variables and then nuke the file as soon as they are not used anymore.