add mux docs; allow to specify mux namespaces
This adds the necessary documentation for the mux parameters and
behavior.  This also adds new parameters which allow to specify
the namespaces mux must create by default, and which can be
optionally added.
This casts openshift_logging_use_mux to bool wherever it is used
as a boolean.
Also - use oc_service `labels` and `external_ips` which are now
available.